feat: 🎸 add Admin API and big refactor

+++ src/services/book.service.ts
@@ -1,19 +1,24 @@
 import { Injectable } from '@nestjs/common'
+import config from '@root/app/config/appConfig'
+
+import { BadRequestException, UnauthorizedException } from '@root/app/exception/httpException'
+import { httpFlags } from '@root/constant/flags'
+
 import { IBook } from '@root/database/models/book.model'
+import { UpdateBookEnum, BookStatusEnum } from '@root/interfaces/enum'
+
 import { BookRepository } from '@root/repositories/book.repository'
 import { UserRepository } from '@root/repositories/user.repository'
 import { CategoryRepository } from '@root/repositories/category.repository'
-import { BadRequestException } from '@root/app/exception/httpException'
-import { httpFlags } from '@root/constant/flags'
-import { UpdateBookEnum, BookStatusEnum } from '@root/interfaces/enum'
-import config from '@root/app/config/appConfig'
 
+import { UserService } from '@root/services/user.service'
 @Injectable()
 export class BookService {
   constructor(
     private readonly bookRepository: BookRepository,
     private readonly userRepository: UserRepository,
-    private readonly categoryRepository: CategoryRepository
+    private readonly categoryRepository: CategoryRepository,
+    private readonly userService: UserService
   ) {}
 
   public async findAllBook(page: number) {
@@ -64,7 +69,7 @@ export class BookService {
     body.status = BookStatusEnum.WAIT
     body.uploadBy = userId
 
-    const user = await this.userRepository.getOneUser(userId)
+    const user = await this.userService.findOneUser(userId)
     const count = user.uploadedBook.length || 0
 
     body.file = 'file-' + userId + `-${count}.pdf`
@@ -135,13 +140,21 @@ export class BookService {
     }
   }
 
-  public async deleteBook(id: string) {
+  public async deleteBook(id: string, userId: string) {
     const book = await this.bookRepository.getOneBook(id, false)
+    const user = await this.userService.findOneUser(userId)
     if (book) {
-      book.isActive ? await this.updateBookOnCategory(book.categoryIds, book._id, UpdateBookEnum.DELETE) : null
-      await this.bookRepository.deleteOneBook(id)
-      return {
-        message: `Book with id: ${id} has successfully deleted`
+      if (user.isAdmin || book.uploadBy.toString() === userId) {
+        book.isActive ? await this.updateBookOnCategory(book.categoryIds, book._id, UpdateBookEnum.DELETE) : null
+        await this.bookRepository.deleteOneBook(id)
+        return {
+          message: `Book with id: ${id} has successfully deleted`
+        }
+      } else {
+        throw new UnauthorizedException(
+          httpFlags.UNAUTHORIZED,
+          `You don't have permission to delete book with id: ${id}`
+        )
       }
     } else {
       throw new BadRequestException(httpFlags.BOOK_NOT_FOUND)
